chore(common): CHECKOUT-6970 Move form provider to context package
