+++ packages/locale/src/index.ts
@@ -1,2 +1,2 @@
 export { default as TranslatedString } from './TranslatedString/TranslatedString';
-export { LocaleContext, LocaleContextType, useLocale } from './context';
+export { LocaleContext, LocaleContextType, useLocale } from './contexts';
