--- packages/payment-integration-api/src/providers/index.ts
@@ -1,1 +1,0 @@
-export { default as FormProvider } from './FormProvider';
