+++ packages/ui/src/form/contexts/index.ts
@@ -1,1 +1,1 @@
-export { default as FormContext, FormContextType } from './FormContext';
+export { default as FormContext, FormContextType, FormProvider } from './FormContext';
